clean up edgecon
edgecon currently aims to output a filled in shape for each z layer.
main conditions for creating lines: points on different segments
initially, and points are found as endpoints
issues:  unable to form a closed shape: if you have a U shape from one
of the earlier iterations, it is all one segment, so it won't connect
the top. Also, if one end is rounded, or if trying to connect to a small
closed shape, will be unable to because it won't find any endpoints on
the circle or whatever.

for a good example run and look at layer z = 7.  there is only one
endpoint near the opening, so it can't connect to anything.