worked on edgecon
now have 2 'directions' for endpoints, the gradient of the ct scan and
the direction of the traced line.
each of these is called in a function (lbldir for direction from lines;
ctdir for gradient direction (which uses imggrad function in it)).

direction was renamed to lbldir. testfile.txt.txt was deleted because
there was nothing in it.